+++ SP_3D_1ceng/initialPML.m
@@ -1,0 +1,18 @@
+function pml = initialPML(nodeCoordinates,gamma1_2,pitch,msh,material_type)
+pml_dmax = 1e11;
+pml_n = 3;
+
+pml_xp_low = nodeCoordinates(gamma1_2(1),3);   % 底部厚度     多层结构导致底部xa不等于0 需要设置初始值
+pml_xa_low = pml_xp_low+2*pitch;   % 底部厚度
+
+pml(3).indx = find(msh.HEXAS27(:,28) == 9)';
+
+pml(3).za = pml_xa_low;
+pml(3).zp = pml_xp_low;
+pml(3).dmax = pml_dmax;
+pml(3).n = pml_n;
+pml(3).rho = material_type(1).rho;
+pml(3).C = material_type(1).C;
+pml(3).e = material_type(1).e;
+pml(3).eps = material_type(1).eps;
+end