+++ codes/AssemblyKM3D.m
@@ -1,0 +1,6 @@
+[massuu,stiffnessuu,stiffnessfifi,stiffnessufi,stiffnessfiu] = formStiffnessMass3D(GDofu,GDoffi,...
+    elementNodes,numberNodes,nodeCoordinates,material_type);
+
+omega = 2*pi*fre(end);
+[mass,stiffness] = formStiffnessMass3DPML(massuu,stiffnessuu,stiffnessfifi,stiffnessufi,stiffnessfiu,...
+    elementNodes,numberNodes,nodeCoordinates,pml,omega);